backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
httpx<0.28
langgraph==0.2.13
langchain==0.2.13
langchain-openai==0.2.13
langchain-community==0.2.13
